+++ .storybook-vnext/preview.js
@@ -5,6 +5,7 @@ import '../css/tokens/generated/index.css';
 import '../css/components/accordion.css';
 import '../css/components/alert.css';
 import '../css/components/badge.css';
+import '../css/components/breadcrumb.css';
 import '../css/components/button.css';
 import '../css/components/card.css';
 import '../css/components/checkbox.css';